requirements.txt
bcrypt
beautifulsoup4
blinker
blis
catalogue
certifi
cffi
chardet
ChatterBot
click
cymem
elasticsearch
email_validator
Flask
Flask-Admin
Flask-Bcrypt
Flask-Login
Flask-Mail
Flask-SQLAlchemy
Flask-WTF
html5lib
idna
importlib-metadata
itsdangerous
Jinja2
joblib
lxml
MarkupSafe
marshmallow
mathparse
murmurhash
nltk
numpy
pandas
pbr
Pillow
Pint
plac
preshed
pycparser
python-dateutil
pytz
PyYAML
regex
requests
scikit-learn
scipy
six
sklearn
soupsieve
spacy
SQLAlchemy
srsly
thinc
threadpoolctl
tqdm
urllib3
wasabi
webargs
webencodings
Werkzeug
Whoosh
WTForms
zipp
psycopg2-binary
